Change IdentityBlockl::items from std::vector to QList

+++ cryptutil.cpp
@@ -276,7 +276,7 @@ bool CryptUtil::decryptBlock1(QByteArray& decryptedImk, QByteArray& decryptedIlk
     QByteArray verificationTag = QByteArray::fromHex(block->items[13].value.toLocal8Bit());
     QByteArray encryptedIdentityKeys = encryptedImk + encryptedIlk;
     QByteArray plainText;
-    for (size_t i=0; i<11; i++) plainText.append(block->items[i].toByteArray());
+    for (int i=0; i<11; i++) plainText.append(block->items[i].toByteArray());
 
     int ret = crypto_aead_aes256gcm_decrypt_detached(
                 reinterpret_cast<unsigned char*>(decryptedIdentityKeys.data()),
@@ -324,7 +324,7 @@ bool CryptUtil::decryptBlock2(QByteArray &decryptedIuk, IdentityBlock *block, QS
     QByteArray encryptedIuk = QByteArray::fromHex(block->items[5].value.toLocal8Bit());
     QByteArray verificationTag = QByteArray::fromHex(block->items[6].value.toLocal8Bit());
     QByteArray plainText;
-    for (size_t i=0; i<5; i++) plainText.append(block->items[i].toByteArray());
+    for (int i=0; i<5; i++) plainText.append(block->items[i].toByteArray());
 
     QByteArray key;
     bool ok = CryptUtil::enScryptIterations(
@@ -380,11 +380,11 @@ bool CryptUtil::decryptBlock3(QList<QByteArray> &decryptedPreviousIuks, Identity
             );
 
     QByteArray plainText;
-    for (size_t i=0; i<3; i++) plainText.append(block->items[i].toByteArray());
+    for (int i=0; i<3; i++) plainText.append(block->items[i].toByteArray());
 
     QByteArray encryptedData;
 
-    for (size_t i=0; i<static_cast<size_t>(nrOfPreviousIuks); i++)
+    for (int i=0; i<static_cast<int>(nrOfPreviousIuks); i++)
     {
         encryptedData.append(
                     QByteArray::fromHex(block->items[3+i].value.toLocal8Bit())
@@ -677,7 +677,7 @@ IdentityBlock CryptUtil::createBlock1(QByteArray iuk, QString password,
     // Encrypt identity keys
     QByteArray unencryptedKeys = imk + ilk;
     QByteArray additionalData;
-    for (size_t i=0; i<11; i++) additionalData.append(block1.items[i].toByteArray());
+    for (int i=0; i<11; i++) additionalData.append(block1.items[i].toByteArray());
     QByteArray encryptedData = aesGcmEncrypt(unencryptedKeys, additionalData, initVec, key);
     QByteArray encryptedImk = encryptedData.left(32);
     QByteArray encryptedIlk = encryptedData.mid(32, 32);
@@ -724,7 +724,7 @@ IdentityBlock CryptUtil::createBlock2(QByteArray iuk, QString rescueCode, QProgr
     block2.items[4].value = QString::number(iterationCount);
 
     // Encrypt IUK
-    for (size_t i=0; i<5; i++) additionalData.append(block2.items[i].toByteArray());
+    for (int i=0; i<5; i++) additionalData.append(block2.items[i].toByteArray());
     QByteArray encryptedData = aesGcmEncrypt(iuk, additionalData, initVec, key);
     QByteArray encryptedIuk = encryptedData.left(32);
     QByteArray authTag = encryptedData.right(16);
@@ -842,7 +842,7 @@ bool CryptUtil::updateBlock1(IdentityBlock *block1, QByteArray unencryptedImk,
     QByteArray unencryptedKeys = unencryptedImk + unencryptedIlk;
     QByteArray encryptedKeys(unencryptedKeys.length(), 0);
     QByteArray authTag(16, 0);
-    for (size_t i=0; i<11; i++) newPlainText.append(block1->items[i].toByteArray());
+    for (int i=0; i<11; i++) newPlainText.append(block1->items[i].toByteArray());
     unsigned long long authTagLen;
 
     crypto_aead_aes256gcm_encrypt_detached(
